🐛 [MODIFY](custom): query 로직 수정
