+++ src/api/index.ts
@@ -1,0 +1,10 @@
+import axios from "axios";
+
+const config = {
+  //   baseURL: "",
+  //   withCredentials: true,
+};
+
+const api = axios.create(config);
+
+export default api;
