+++ src/hooks/query/useGlobalQuery.tsx
@@ -1,11 +1,11 @@
-import axios from "axios";
-import useQuery from "@/hooks/useQuery";
+import api from "@/api";
+import { useQuery } from "@tanstack/react-query";
 import { queryInformationType } from "query";
 
 const useGlobalQuery = ({ key, URL }: queryInformationType) => {
   const { data, isError, isSuccess, error } = useQuery({
     queryKey: [key, URL],
-    queryFn: () => axios.get(URL),
+    queryFn: () => api.get(URL),
   });
 
   return { data, isError, isSuccess, error };
