--- src/http/index.ts
@@ -1,10 +1,0 @@
-import axios from "axios";
-
-const config = {
-  baseURL: "",
-  withCredentials: true,
-};
-
-const http = axios.create(config);
-
-export default http;
