+++ src/services/request/getRequest.ts
@@ -1,0 +1,4 @@
+export const GET_GLOBAL_REQUEST = {
+  key: "test_data_2",
+  URL: "https://api.github.com/repos/tannerlinsley/react-query",
+};
